Changelog for 2.6.0.

+++ iOS/samples/AVOSDemo/AVOSDemo/AVFileBasic.m
@@ -77,7 +77,7 @@
                     [self log:[NSString stringWithFormat:@"成功得到图片: %@",[img description]]];
                 }
             } progressBlock:^(NSInteger percentDone) {
-                [self log:[NSString stringWithFormat:@"加载进度: %d%%",percentDone]];
+                [self log:[NSString stringWithFormat:@"加载进度: %ld%%", (long)percentDone]];
             }];
             
         }else{
